dashboard frontend version 2.0 successfully integrated and tested with the backend API
