+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2025_7_5_14_21__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2025_7_5_14_21_[Changes]" date="1751721671865" recycled="false" toDelete="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2025_7_5_14_21_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2025/7/5 14:21 [Changes]" />
+</changelist>
