+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2025_7_6_11_54__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2025_7_6_11_54_[Changes]" date="1751799283491" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2025_7_6_11_54_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2025/7/6 11:54 [Changes]" />
+</changelist>
