+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2025_7_6_11_55__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2025_7_6_11_55_[Changes]" date="1751799313163" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2025_7_6_11_55_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2025/7/6 11:55 [Changes]" />
+</changelist>
