+++ frontend/urbanflow_frontend/src/views/dashboard/TrafficFlowChart.vue
@@ -9,8 +9,8 @@ import { CanvasRenderer } from 'echarts/renderers'
 import { LineChart } from 'echarts/charts'
 import { TitleComponent, TooltipComponent, GridComponent } from 'echarts/components'
 import VChart from 'vue-echarts'
-import { getTrafficFlow } from '@/mocks/mockDashboardData'
-// import { getTrafficFlow } from '@/service/dashboard_api'
+//import { getTrafficFlow } from '@/mocks/mockDashboardData'
+import { getTrafficFlow } from '@/service/dashboard_api'
 
 use([CanvasRenderer, LineChart, TitleComponent, TooltipComponent, GridComponent]);
 
@@ -65,33 +65,31 @@ async function fetchData() {
 
   const response = await getTrafficFlow(params);
 
-  if (response && response.data && response.labels) {
-    chartOption.value.xAxis.data = response.labels;
-    // NOTE: Backend returns the full data object, we extract the value here.
-    chartOption.value.series[0].data = response.data.map((d: any) => d.flow_rate_hourly);
+//   if (response && response.data && response.labels) {
+//     chartOption.value.xAxis.data = response.labels;
+//     // NOTE: Backend returns the full data object, we extract the value here.
+//     chartOption.value.series[0].data = response.data.map((d: any) => d.flow_rate_hourly);
+//   } else {
+//     chartOption.value.xAxis.data = [];
+//     chartOption.value.series[0].data = [];
+//   }
+
+  if (response && response.data && response.xAxisLabels && response.yAxisConfig) {
+    // 更新X轴标签
+    chartOption.value.xAxis.data = response.xAxisLabels;
+
+    // 更新Y轴配置
+    chartOption.value.yAxis.min = response.yAxisConfig.min;
+    chartOption.value.yAxis.max = response.yAxisConfig.max;
+    chartOption.value.yAxis.interval = response.yAxisConfig.interval;
+
+    // 更新图表数据
+    chartOption.value.series[0].data = response.data.map((d: any) => d.flowRateHourly);
   } else {
+    // 如果接口出错或返回数据不规范，清空图表
     chartOption.value.xAxis.data = [];
     chartOption.value.series[0].data = [];
   }
-
-  // if (response && response.data && response.xAxisLabels && response.yAxisConfig) {
-  //   // 更新X轴标签
-  //   chartOption.value.xAxis.data = response.xAxisLabels;
-  //
-  //   // 更新Y轴配置
-  //   chartOption.value.yAxis.min = response.yAxisConfig.min;
-  //   chartOption.value.yAxis.max = response.yAxisConfig.max;
-  //   chartOption.value.yAxis.interval = response.yAxisConfig.interval;
-  //
-  //   // 更新图表数据
-  //   chartOption.value.series[0].data = response.data.map((d: any) => d.flow_rate_hourly);
-  // } else {
-  //   // 如果接口出错或返回数据不规范，清空图表
-  //   chartOption.value.xAxis.data = [];
-  //   chartOption.value.series[0].data = [];
-  // }
-
-
 }
 
 watch(() => props.filters, fetchData, { deep: true });
